feat(tesseract): Rolling window over symbols
