+++ rust/cubesqlplanner/cubesqlplanner/src/planner/planners/multi_stage/mod.rs
@@ -3,11 +3,9 @@ mod member;
 mod member_query_planner;
 mod multi_stage_query_planner;
 mod query_description;
-mod rolling_window_planner;
 
 pub use applied_state::*;
 pub use member::*;
 pub use member_query_planner::MultiStageMemberQueryPlanner;
 pub use multi_stage_query_planner::MultiStageQueryPlanner;
 pub use query_description::MultiStageQueryDescription;
-pub use rolling_window_planner::RollingWindowPlanner;
